bad_topics update
Updated the code to include a bad words list to check before saving reddit posts. It looks for posts in the titles and then based on that decides as another factor whether to save the post or not. This should hopefully help cut down a chunk of the cancerous posts being saved.