Added Error Checker
Fixed the bot so that if it gets an error return code from a FB post it will log it but try to post a different post again until it doesn't get an error. This will work for now. though the primary issue with this is that FB servers might get spammed if the bot is constantly trying and failing posts. So I may hit a limit on API calls or get the bot banned for that. Luckily in our case, error'd posts are very rare so statistically it is very unlikely. Still... for the sake of good programming I will make it a point to add that in in before next month probably.