added error counter
Might as well.. just went ahead and did it. Added a counter for errors too so if it comes back with an error 10 times in a row it just gives up and stops trying to post until the next interval. This is really an "if all else fails" kind of scenario because it's extremely unlikely currently. Maybe later down the road when the bot gets to have thousands of posts to choose from in the reddit log, then it might become an issue. For now it will work and we can always tweak it, modify its post list and what not. Should be fine though. :)