Update FBPoster-Cloud.py
Updated the code to include dHashing now to account for duplicates! Also updated to log to a spreadsheet for cleaner logging instead of to notepad (idk why I thought that was a good idea originally lol). Big update here! 

Note, the code now has to be run 4 times as much for now since it will only log one post at a time now instead of 4. -- will fix that part later... eventually :p