Update RedditGrabber-Cloud.py
Fixed the duplicate grabbing issue. My previous implementation did not work. 

Here's a summary of the issue and how I fixed it: basically I was hashing the images right but when I was searching for the hash strings within the spreadsheet, it wasn't finding any duplicates. at first I thought this worked, but upon further investigation I was wrong. The issue was that my method of searching a string in a list of lists did not work. I had to first flatten the list and then search within that. Because all we're really doing is saying if string in list then discard and try a different one, but if not then we're good. 

I also had to adjust my bad words list code too because for similar reasons it also didn't work. I also re increased the grabbing back to 4 each time so now we're back in business there.