Update FBPoster-Cloud.py
TLDR; I had to flatten the list of reddit grabber spreadsheet item list so that we can search strings through it. Basically the code will pick a random "list" (row) within the reddit grabber list of lists (spreadsheet) and it will check the image hash of that list (last item within the list it picked) to see if it is already in the fb poster list of lists (spreadsheet). if it is then that means it has posted it before so it will remove it. I thought I had this working before but turns out I had to flatten the list of lists into just a single list to search the string through it. 

TODO: This code is not optimized well for scalability. The FB Poster needs to include a way to remove the row of whatever list it grabbed from reddit spreadsheet. I can't think of a clean way to do this right now but it might be a simple fix. if we can pull that off then the reddit grabber list won't be full of redundant posts that don't need to be there once posted.