Update FBPoster-Cloud.py
Updated the code so that every time it grabs from the reddit spreadsheet and posts to FB, it will remove that row from the reddit spreadsheet. (this will help keep compute times lower and reduce the redundancy in the reddit spreadsheet. This way the reddit spreadsheet only serves as a buffer for possible posts but doesn't store all the same data over time. If we didn't do this, it would have to compare a 2d list of hundreds of thousands of items to another 2d list of hundreds of thousands of items. 

Theoretically now that it removes the item after posting, we can get rid of the check hash function and cut down on a ton of time. But I'll remove that later I think. We'll keep it for the time being. :)