Duplicate post update (again)
Apparently I'm a moron and did not have Reddit Grabber check to make sure the image hash was not in the RG spreadsheet. It was checking FB Poster but not RG too, which is dumb. But at least FB Poster was smart enough to not post duplicates and that all that really matters to end users. 

Anyways, RG script should be fixed now. I'm getting a Raspberry Pi 4 soon that I will host this script on going forward, when that comes, I will implement OCR into the code again so we can filter out the bad memes using the text within the images. Excited for that update!