OCR Implemented (and switched RPi4 instead of AWS)
In this update I have switched from my previous implementation I was using to test (EasyOCR) to Tesseract OCR. I prefer EasyOCR for usability but unfortunately EasyOCR requires PyTorch dependencies which are not supported on ARM CPUs. So whatever I guess. Wasted 2 straight days trying to finally figure that out. (I'm a moron lol). But made the switch to Tesseract and so far it's been much better and easier to set up on this particular system. So yay! Now we're back in business! 

Also switched the bot off AWS onto Raspberry Pi because quite frankly OCR is too hefty to run on a free AWS server (currently) and it would just be cheaper to make a one time payment of a raspberry pi + monthly electricity bill of RPi4 instead of the money it would require to get the same specs in an AWS server every single month. But I have a feeling that eventually I'll migrate back over one day if we can generate enough revenue / donations to make it work. For now though, pi it is!