adding captions to posts
Been wanting to do this for a while but wasn't sure how, turned out to be remarkably easy. I originally wanted to make a post with a message included but still haven't figured out how to do that, but this is the next best thing!. Though it does double our API calls which kinda sucks, but luckily we only post once an hour so it's really just moving from 1 api call an hour to 2 so not really that big of a deal.