Fixed the formatting so it can collect 4 posts now
The bot will now pick a random subreddit, collect 4 posts that meet the criteria, then stop. In this instance it is pulling more from each source when it picks a source. This does result in less variation overall (in the instance for example it picks a subreddit A, grabs 4 posts from A, then the next time it runs it ends up picking A again and grab 4 more posts, that's 8 posts grabbed in 4 hours from a single source). But overall randomness will eventually level this out. If you do the math long term it doesn't have any major long term repercussions doing it this way. Mainly the reason I wanted to do this is so that we don't have to run the code so frequently. Now it only has to run every 2 hours instead of every 30 minutes. Which lessens the chances of failure and helps it build up a queue if for some reason something goes down.