Fixed syntax error in 112
the IF clause was supposed to be "If Page ID **IN** return_text_str" not "If Page ID not in return_text_str". This led to an error where it would make a FB post but would not log it to FB logger or edit the caption. I knew exactly where the error was because if you follow the branches you can see exactly where it stops. Silly mistake I made in a huge amount of refactoring so it's whatevs.